+++ src/scriptlets/scriptlets-names-list.ts
@@ -77,6 +77,7 @@ export { preventCanvasNames } from './prevent-canvas';
 export { trustedReplaceArgumentNames } from './trusted-replace-argument';
 export { preventInnerHTMLNames } from './prevent-innerHTML';
 export { preventConstructorNames } from './prevent-constructor';
+export { removeRequestQueryParameterNames } from './remove-request-query-parameter';
 // redirects as scriptlets
 // https://github.com/AdguardTeam/Scriptlets/issues/300
 export { AmazonApstagNames } from './amazon-apstag';